backend/foodgram/requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.9.3
Django==4.0.4
djangorestframework==3.13.1
django-filter==21.1
python-dotenv==0.20.0
djoser==2.1.0
